refactored trusted part of OFTonSGX, I have broken the huge files enclave.c and enclave.edl into several different files
