+++ OFTonSGX/trusted/stats/stats.edl
@@ -1,0 +1,26 @@
+enclave {
+	include "common.h"
+    trusted {
+
+		// Vanilla ECALLS
+
+        public size_t
+        ecall_flow_stats_c(uint8_t bridge_id);
+        public size_t
+        ecall_flow_stats_r(uint8_t bridge_id,[user_check]struct cls_rule **buf,int elem);
+
+		// Optimized ECALLS
+
+		public size_t
+		ecall_collect_rules_loose_stats_request(uint8_t bridge_id,
+											  uint8_t table_id,
+											  int n_tables,
+											  size_t start_index,
+											  size_t buffer_size,
+											  [user_check] struct match *match,
+											  [user_check] struct cls_rule **cls_rules,
+											  [user_check] struct match *matches,
+											  [user_check] unsigned int *priorities,
+											  [user_check] size_t *n_rules);
+    };
+};
